+++ src/uninstall.js
@@ -4,5 +4,6 @@ import defs from './utils/defs'
 import { exists, rmdir } from './utils/fs'
 
 export function* completion(templateName) {
+    [templateName] = templateName.split(defs.versionSep)
     yield rmdir(`${defs.defaults.pkgPath}${sep}${templateName}`)
 }
